+++ internal/oauth/store/capability_store_test.go
@@ -1,4 +1,4 @@
-package aggregator
+package store
 
 import (
 	"context"